requirements.txt
matplotlib
opencv-python
scikit-image
numpy
tensorboardX
torch
PyYAML
transforms3d
easydict
cffi

